modularize the custom ssg
