+++ scripts/static_site_builder/__init__.py
@@ -1,0 +1,3 @@
+"""Helpers for the static Digital Giza site builder."""
+
+"""Helpers for the static Digital Giza site builder."""
