+++ templates/static-site/footer.html
@@ -1,0 +1,28 @@
+<nav class="footer">
+  <div class="row">
+    <section class="large-6 columns">
+      <div class="row">
+        <div class="medium-4 columns"><h6><a class="heading-link" href="/library/">Library</a></h6></div>
+        <div class="medium-4 columns"><h6><a class="heading-link" href="/gizaatschool/">Giza @ School</a></h6><ul class="footer-menu-list"><li><a href="/lessons/">Lesson Topics</a></li><li><a href="/commontopics/">People and Places of Giza</a></li><li><a href="/faq/">Glossary and FAQ</a></li><li><a href="/videos/">Video Library</a></li></ul></div>
+        <div class="medium-4 columns"><h6><a class="heading-link" href="/giza3d/">Giza 3D</a></h6><ul class="footer-menu-list vertical menu accordion-slim" data-accordion-menu><li><a href="#0" class="accordion-toggletext" data-closed-text="Show models" data-open-text="Hide models">Show models</a><ul class="no-bullet menu vertical nested"><li><a href="/giza3d/">Giza Plateau</a></li><li><a href="/giza3d/?mode=GuidedTour&amp;guidedTourId=neZZCDSf5">A Walking Tour of the Giza Plateau</a></li><li><a href="/giza3d/?mode=GuidedTour&amp;guidedTourId=neZZCDSf5">Khafre Pyramid and Temples</a></li><li><a href="/giza3d/?mode=GuidedTour&amp;guidedTourId=neZZCDSf5">Khufu Pyramid</a></li><li><a href="/giza3d/?mode=GuidedTour&amp;guidedTourId=neZZCDSf5">Menkaure Pyramid</a></li><li><a href="/giza3d/?mode=GuidedTour&amp;guidedTourId=neZZCDSf5">Great Sphinx</a></li><li><a href="/giza3d/?mode=GuidedTour&amp;guidedTourId=nGQi15GpZ">Great Sphinx Interactive Video</a></li></ul></li></ul></div>
+      </div>
+    </section>
+    <section class="large-6 columns">
+      <div class="row">
+        <div class="medium-4 columns"><h6><a class="heading-link" href="/collections/">My Giza</a></h6><ul class="footer-menu-list"><li><a href="/collections/">Browse Collections</a></li></ul></div>
+        <div class="medium-4 columns"><h6><a class="heading-link" href="/gizaintro/">About the Project</a></h6><ul class="footer-menu-list"><li><a href="/gizaintro/">Introduction to Giza</a></li><li><a href="/about/">What is the Giza Project?</a></li><li><a href="/archaeology/">Archaeology at Giza</a></li></ul></div>
+        <div class="medium-4 columns"><h6><a class="heading-link" href="https://community.alumni.harvard.edu/give/34086571">Donate</a></h6><h6><a class="heading-link" href="/search/">Search</a></h6><h6><a class="heading-link" href="/contact/">Contact</a></h6></div>
+      </div>
+    </section>
+  </div>
+</nav>
+<footer class="legal-social" data-equalizer data-equalize-on="medium">
+  <div class="row">
+    <div class="large-6 large-push-6 columns text-md-right" data-equalizer-watch>
+      <img src="/static/images/footerlogo.png" alt="Giza Project at Harvard University" class="image-spacer" style="max-width:200px;">
+      <img src="/static/images/neh_logo_web.svg" alt="National Endowment for the Humanities" class="image-spacer" style="margin-top:-0.2em; max-width:200px;">
+      <p class="medium-9 columns text-smaller text-muted text-center">Digital Giza has been made possible in part by the <a href="http://www.neh.gov/">National Endowment for the Humanities</a>: Exploring the human endeavor</p>
+    </div>
+    <div class="large-6 large-pull-6 columns" data-equalizer-watch><p class="text-smaller text-muted" style="margin-top:0.25em;">&copy; 2026 Giza Project at Harvard University. <br> All rights reserved.</p></div>
+  </div>
+</footer>
