+++ templates/static-site/header.html
@@ -1,0 +1,38 @@
+<a id="skippy" class="sr-only sr-only-focusable" href="#content">
+  <div class="top-bar-container"><span class="skiplink-text">Skip to main content</span></div>
+</a>
+<div data-sticky-container>
+  <div data-sticky data-options="marginTop:0;" style="width:100%; z-index:1003;">
+    <div class="top-bar-container" style="width:100%;">
+      <div class="row row-padded">
+        <div class="title-bar" data-responsive-toggle="main-menu" data-hide-for="medium">
+          <button class="icon-bars icon-padded title-bar-title" type="button" data-toggle>Menu</button>
+        </div>
+        <div class="top-bar" id="main-menu">
+          <div class="top-bar-left hide-for-small-only"><a href="/"><img src="/static/images/navlogo.png" alt="Digital Giza at Harvard University. Click for home."></a></div>
+          <div class="top-bar-left nav-menu-primary">
+            <ul class="vertical medium-horizontal menu" data-responsive-menu="drilldown medium-dropdown" data-alignment="left">
+              <li><a href="/library/">Library</a></li>
+              <li><a href="/gizaatschool/">Giza @ School</a></li>
+              <li><a href="/giza3d/">Giza 3D</a></li>
+              <li><a href="/collections/">My Giza</a>
+                <ul class="vertical menu">
+                  <li><a href="/collections/">Browse Collections</a></li>
+                </ul>
+              </li>
+              <li><a href="/gizaintro/">About</a>
+                <ul class="vertical menu">
+                  <li><a href="/gizaintro/">Introduction to Giza</a></li>
+                  <li><a href="/about/">What is the Giza Project?</a></li>
+                  <li><a href="/archaeology/">Archaeology at Giza</a></li>
+                </ul>
+              </li>
+              <li><a href="https://community.alumni.harvard.edu/give/34086571">Donate</a></li>
+              <li><form id="search-form" action="/search-results/"><input type="text" id="inputSimpleSearch" name="q" placeholder="Search"><button class="icon-search button" type="submit"><span class="show-for-sr">Search</span></button></form></li>
+            </ul>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
