+++ templates/static-site/home.html
@@ -1,0 +1,60 @@
+<section class="home-hero-container">
+  <div class="row row-padded">
+    <div class="video-bg-home">
+      <div class="static-site-home-video">
+        <video autoplay loop muted playsinline poster="/static/video/home/GizaHome_web.jpg">
+          <source src="/static/video/home/GizaHome_web.mp4" type="video/mp4">
+          <source src="/static/video/home/GizaHome_web.webm" type="video/webm">
+          <source src="/static/video/home/GizaHome_web.ogv" type="video/ogg">
+        </video>
+      </div>
+      <div class="home-hero-title"><h1><span class="title-text-alt">Welcome to the</span> Giza Plateau</h1></div>
+      <div class="home-hero-content">
+        <div class="medium-8 columns">
+          <p class="lead text-alt" style="line-height:1.4">The Giza Project gives you access to the largest collection of information, media, and research materials ever assembled about the Pyramids and related sites on Egypt&rsquo;s Giza Plateau.</p>
+        </div>
+        <div class="medium-4 columns">
+          <form action="/search-results/">
+            <p class="text-alt" style="color:#fff; margin-bottom:0.1rem;"><em>Search the archives:</em></p>
+            <div class="input-group static-site-hero-search">
+              <input class="input-group-field" type="text" name="q" placeholder="Search">
+              <div class="input-group-button"><button type="submit" class="button icon-search"><span class="show-for-sr">Search</span></button></div>
+            </div>
+          </form>
+          <small>or go to <a class="link-lighter" href="/search/">Advanced Search</a></small>
+        </div>
+      </div>
+      <div class="home-hero-overlay"></div>
+    </div>
+  </div>
+</section>
+<div class="row p-r-1 static-site-home-features" data-equalizer data-equalize-on="medium">
+  <section class="medium-6 large-3 columns">
+    <div class="d-flex flex-v border-left-dark p-l-1 m-r-neg1 m-b-2" data-equalizer-watch>
+      <div class="flex-header"><h2 class="h2-minor text-medium m-b-1"><a class="heading-link" href="/giza3d/">Explore Giza 3D</a></h2></div>
+      <div class="flex-body text-smaller"><p><a href="/giza3d/"><img class="img-fluid img-fluid-mw325 border-dark" src="/static/images/home-feature-giza3d.png" alt="Digital rendering of the Giza Plateau as seen from the air"></a></p><p>Immerse yourself in <strong>realistic 3D reconstructions</strong> of the Giza plateau.</p></div>
+      <div class="flex-footer"><div class="row"><div class="columns small-12 md-phone-6 medium-12 lg-tablet-8 large-12"><a href="/giza3d/" class="button primary expanded m-b-0">Jump In <i class="icon-angle-right"></i></a></div></div></div>
+    </div>
+  </section>
+  <section class="medium-6 large-3 columns">
+    <div class="d-flex flex-v border-left-dark p-l-1 m-r-neg1 m-b-2" data-equalizer-watch>
+      <div class="flex-header"><h2 class="h2-minor text-medium m-b-1"><a class="heading-link" href="/gizaatschool/">Giza @ School</a></h2></div>
+      <div class="flex-body text-smaller"><p><a href="/gizaatschool/"><img class="img-fluid img-fluid-mw325 border-dark" src="/static/images/home-feature-school.png" alt="Digital rendering of a figure from Giza's past standing in front of a tomb entrance"></a></p><p>Resources designed especially for <strong>teachers and students</strong>.</p></div>
+      <div class="flex-footer"><div class="row"><div class="columns small-12 md-phone-6 medium-12 lg-tablet-8 large-12"><a href="/gizaatschool/" class="button primary expanded m-b-0">View Resources <i class="icon-angle-right"></i></a></div></div></div>
+    </div>
+  </section>
+  <section class="medium-6 large-3 columns">
+    <div class="d-flex flex-v border-left-dark p-l-1 m-r-neg1 m-b-2" data-equalizer-watch>
+      <div class="flex-header"><h2 class="h2-minor text-medium m-b-1"><a class="heading-link" href="/collections/">Make it Yours</a></h2></div>
+      <div class="flex-body text-smaller"><p><a href="/collections/"><img class="img-fluid img-fluid-mw325 border-dark" src="/static/images/home-feature-mygiza.png" alt="Digital rendering of painted wall decoration"></a></p><p>Browse public collections of related Digital Giza records.</p></div>
+      <div class="flex-footer"><div class="row"><div class="columns small-12 md-phone-6 medium-12 lg-tablet-8 large-12"><a href="/collections/" class="button primary expanded m-b-0">Browse Collections <i class="icon-angle-right"></i></a></div></div></div>
+    </div>
+  </section>
+  <section class="medium-6 large-3 columns">
+    <div class="d-flex flex-v border-left-dark p-l-1 m-r-neg1 m-b-2" data-equalizer-watch>
+      <div class="flex-header"><h2 class="h2-minor text-medium m-b-1"><a class="heading-link" href="https://community.alumni.harvard.edu/give/34086571">Support the Project</a></h2></div>
+      <div class="flex-body text-smaller"><p>The Giza Project, an international collaboration based at Harvard University, aims to <strong>assemble and provide access to all archaeological records</strong> about the most famous site in the world: the Pyramids, surrounding cemeteries and settlements of Giza, Egypt.</p></div>
+      <div class="flex-footer"><div class="row"><div class="columns small-12 md-phone-6 medium-12 lg-tablet-8 large-12"><a href="https://community.alumni.harvard.edu/give/34086571" class="button primary expanded m-b-0">Donate Now <i class="icon-angle-right"></i></a></div></div></div>
+    </div>
+  </section>
+</div>
