+++ templates/static-site/layout.html
@@ -1,0 +1,28 @@
+<!doctype html>
+<html class="no-js" lang="en">
+<head>
+  <meta charset="utf-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <title>Digital Giza | {{ title }}</title>
+  <meta name="description" content="{{ description }}">
+  <link rel="shortcut icon" href="/static/images/favicon.ico">
+  <link href="https://fonts.googleapis.com/css?family=Montserrat:400,500,600" rel="stylesheet">
+  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.3/css/all.css" integrity="sha384-UHRtZLI+pbxtHCWp1t77Bi1L4ZtiqrqD80Kn4Z8NTSRyMA2Fd33n5dQ8lWUE00s/" crossorigin="anonymous">
+  <link rel="stylesheet" href="/static/css/app.css">
+  <link rel="stylesheet" href="/static/css/project.css">
+  <link rel="stylesheet" href="/static/static-site/static-site.css">
+  {{ extra_head_html }}
+</head>
+<body class="{{ body_class }}">
+  {{ header_html }}
+  <main id="content"{{ main_attr_html }}>
+{{ body_html }}
+  </main>
+  {{ footer_html }}
+  <script src="/static/js/app.js"></script>
+  <script src="/static/js/giza.js"></script>
+  <script src="/static/static-site/static-site.js"></script>
+  {{ extra_scripts_html }}
+</body>
+</html>
