+++ templates/static-site/search-results.html
@@ -1,0 +1,9 @@
+<div class="page-header header-bg-5"><div class="row title"><header class="large-12 columns"><h1>Search the Archives</h1></header></div></div>
+<div class="row static-site-search-page" data-giza-search-page>
+  <section class="medium-9 medium-push-3 columns">
+    <giza-search-results class="static-site-search-results" instance="giza-search" page-size="20"></giza-search-results>
+  </section>
+  <aside class="medium-3 medium-pull-9 columns">
+    <giza-search-sidebar instance="giza-search"></giza-search-sidebar>
+  </aside>
+</div>
