+++ templates/static-site/search.html
@@ -1,0 +1,64 @@
+<div class="page-header header-bg-5"><div class="row title"><header class="large-12 columns"><h1>Search the Archives</h1></header></div></div>
+<div class="row static-site-search-page" data-giza-search-page>
+  <section class="large-12 columns">
+    <section class="static-site-advanced-search" data-pagefind-ignore>
+      <h3>Advanced Search</h3>
+      <p>Narrow your search results using the advanced search fields below.</p>
+      <p class="text-smaller">Need help with abbreviations? Check out our <a href="/faq/#abbr">list of abbreviations on the FAQ page.</a></p>
+
+      <form id="advanced-search-form" action="/search-results/" method="get">
+        <div class="row">
+          <fieldset id="category-radio-selector" class="medium-4 medium-push-8 columns">
+            <legend>Search within the selected category:</legend>
+            <div class="feature-block tertiary">
+              <input id="objects" data-for-id="section-objects" value="objects" type="radio" name="category" checked><label for="objects">Objects</label><br>
+              <input id="tombs" data-for-id="section-sites" value="sites" type="radio" name="category"><label for="tombs">Tombs and Monuments</label><br>
+              <input id="diary" data-for-id="section-objects" value="diarypages" type="radio" name="category"><label for="diary">Diary Pages</label><br>
+              <input id="unpubdocs" data-for-id="section-objects" value="unpubdocs" type="radio" name="category"><label for="unpubdocs">Unpublished Documents</label><br>
+              <input id="mapsandplans" data-for-id="section-objects" value="mapsandplans" type="radio" name="category"><label for="mapsandplans">Maps and Plans</label>
+            </div>
+          </fieldset>
+
+          <section id="section-objects" class="medium-8 medium-pull-4 columns category-section">
+            <label>Object title or name:
+              <input name="objects_title" type="text" placeholder="Example: 'royal statue' or 'Khafre'">
+            </label>
+            <label>Object or ID number:
+              <input name="objects_allnumbers" type="text" placeholder="Example: 'HUMFA_14-11-206'">
+            </label>
+            <label>Material:
+              <input name="objects_medium" type="text" placeholder="Example: 'alabaster'">
+            </label>
+            <label>Findspot:
+              <input name="objects_provenance" type="text" placeholder="Example: 'G 2110'">
+            </label>
+            <label>Date of register entry:
+              <input name="objects_entrydate" type="text" placeholder="Example: 05/23/1924">
+            </label>
+          </section>
+
+          <section id="section-sites" class="medium-8 medium-pull-4 columns category-section" style="display: none;">
+            <label>Tomb/Monument number:
+              <input name="sites_number" type="text" placeholder="Example: 'G 7000 X'">
+            </label>
+            <label>Tomb/Monument name:
+              <input name="sites_sitename" type="text" placeholder="Example: 'Eastern Cemetery'">
+            </label>
+            <label>Site Dates:
+              <input name="sites_datevalues" type="text" placeholder="Example: 'Khufu to Shepseskaf'">
+            </label>
+            <label>People:
+              <input name="sites_people" type="text" placeholder="Example: 'George Reisner'">
+            </label>
+          </section>
+        </div>
+
+        <div class="row">
+          <div class="large-12 columns">
+            <button class="button primary" type="submit">Search</button>
+          </div>
+        </div>
+      </form>
+    </section>
+  </section>
+</div>
